ARM: OMAP3: Add board-anvl
For testing only until we have a real machine ID or the
device tree based booting working properly with PM.

WARNING: For now we are disabling board-omap3beagle.o
from Makefile because of the machine ID conflict. This
hack can be dropped once we have a real machine ID.

Also note that this we cannot merge to mainline tree as
we're moving ARM to use device tree based booting.