anvl: Disable default CPU LED and add pin muxing for it
This disables cpu0 trigger for the GPIO LED. Also let's disable the
input and pull for this pin as it's an output and we have an
external pull.

Note that this does not affect the PCA9632TK LEDs getting cut off
during idle as they are using 2V8_OSC signal from TPS72728YFFT that
we are gating with CLKEN from the PMIC when entering idle. We can
keep this from happening by switching to use ti,twl4030-power-idle
instead of ti,twl4030-power-idle-osc-off at the cost of few extra
mW of power consumption. We currently cannot toggle that during
runtime.