Prepare for evt3 to use gpio_15 for bq24190 interrupt
This should allow validating evt3 hardware for the
charger. As gpio_15 is in gpio bank 1, there is no save
and restore of context needed unlike with gpio_152.

This means we no longer will have a glitch on the
line every time when waking up from off mode.

Note that this patch also assumes that evt3 has an
external pull-up on the interrupt line.