power: bq24190_charger: Check status on PM runtime resume
The Linux generic wakeirq will call bq24190 PM runtime resume on
padconf wakeirq. Let's check the status of the interrupt at that
point. To do that, we need to split the interrupt thread into
two parts.

Note that can cause spurious interrupts on omap3 devices with
bq24190 connected to gpio banks 2 - 5 as there's a glitch on
those pins waking from off mode as listed in "Advisory 1.45".
Devices with this issue should not configure the optional
"wakeup" interrupt in the dts file.