power: bq24190_charger: Fix irq triggering to IRQF_TRIGGER_FALLING
The interrupt signal is TRIGGER_FALLING. This is is specified
in the data sheet "PIN FUNCTIONS":

"The INT pin sends active low, 256 us pulse to hos to report charger
 device status and fault."

Also the direction can be seen in the data sheet "Figure 1. bq24190
with D+/D- Detection and USB On-The-Go (OTG)" that shows a 10k pull-up
resistor installed for the sample configurations.

Fixes: d7bf353fd0aa3 ("bq24190_charger: Add support for TI BQ24190
Battery Charger")
Cc: Mark A. Greer <mgreer@animalcreek.com>
Cc: Matt Ranostay <matt@ranostay.consulting>
Signed-off-by: Liam Breck <kernel@networkimprov.net>
Signed-off-by: Tony Lindgren <tony@atomide.com>